+++ webview/plugin/intellijStylePlugin.ts
@@ -32,8 +32,7 @@ export function intellijStylePostPlugin(): Plugin {
                                         .trim() === '1';
                                     document.documentElement.setAttribute('data-theme', isDarkMode ? 'dark' : 'light');
                                 }
-                            `,
-                            injectTo: 'head'
+                            `
                         },
                         {
                             tag: 'link',
@@ -43,8 +42,7 @@ export function intellijStylePostPlugin(): Plugin {
                                 type: 'text/css',
                                 href: '/intellijStyle.css',
                                 onload: 'onIntellijStyleLoaded()'
-                            },
-                            injectTo: 'head'
+                            }
                         }
                     ]
                 }
@@ -166,8 +164,7 @@ export function intellijStylePrePlugin(): Plugin {
                                 window.zoomOut = zoomOut;
                                 window.actualZoom = actualZoom;
                                 window.reloadStyles = reloadStyles;
-                            `,
-                            injectTo: 'head'
+                            `
                         }
                     ]
                 };